+++ packages/brick_graphql/lib/src/runtime_graphql_definition.dart
@@ -14,7 +14,7 @@ class RuntimeGraphqlDefinition {
 
   /// For fields that are not strictly associations but have nested attributes,
   /// [subfields] needs to be defined for the GraphQL query to resolve.
-  final Set<String> subfields;
+  final Map<String, Map<String, dynamic>> subfields;
 
   /// The type accessed after the result is retrieved, **not** the GraphQL type.
   /// In other words, the runtime type.
@@ -24,7 +24,7 @@ class RuntimeGraphqlDefinition {
     this.association = false,
     required this.documentNodeName,
     this.iterable = false,
-    this.subfields = const <String>{},
+    this.subfields = const <String, Map<String, dynamic>>{},
     required this.type,
   });
 }